服务接口类完成

+++ Generate_Demo/src/main/java/com/java_generate_demo/mappers/BaseMapper.java
@@ -8,32 +8,32 @@ public interface BaseMapper<T, P> {
     /**
      * insert:(插入)
      */
-    Long insert(@Param("bean") T t);
+    Integer insert(@Param("bean") T t);
 
     /**
      * insertOrUpdate:(更新)
      */
-    Long insertOrUpdate(@Param("bean") T t);
+    Integer insertOrUpdate(@Param("bean") T t);
 
     /**
      * delete:(删除)
      */
-    Long delete(@Param("bean") T t);
+    Integer delete(@Param("bean") T t);
 
     /**
      * insertBatch:(批量插入)
      */
-    Long insertBatch(@Param("list") List<T> list);
+    Integer insertBatch(@Param("list") List<T> list);
 
     /**
      * deleteBatch:(批量删除)
      */
-    Long deleteBatch(@Param("list") List<T> list);
+    Integer deleteBatch(@Param("list") List<T> list);
 
     /**
      * insertOrUpdateBatch:(批量插入或更新)
      */
-    Long insertOrUpdateBatch(@Param("list") List<T> list);
+    Integer insertOrUpdateBatch(@Param("list") List<T> list);
 
     /**
      * selectList:(根据参数查询集合)
@@ -43,5 +43,5 @@ public interface BaseMapper<T, P> {
     /**
      * selectCount:(根据集合查询数量)
      */
-    Long selectCount(@Param("query") P p);
+    Integer selectCount(@Param("query") P p);
 }
